Moved isFilePdf to the PdfViewerFragment, renamed auto-generated variables
