+++ app/src/main/res/layout/activity_plans.xml
@@ -12,7 +12,7 @@
 
         <!-- The toolbar aka SupportActionBar -->
         <include layout="@layout/toolbar" />
-    </LinearLayout>LinearLayout>
+    </LinearLayout>
     <!-- The navigation drawer -->
     <include layout="@layout/navigation_drawer" />
